lothitory 22번째줄 null

+++ FinalArduino/Temperature.ino
@@ -3,11 +3,11 @@
 #define DHTTYPE DHT11  // DHT시리즈중 DHT11 선택
 
 DHT dht(DHTPIN, DHTTYPE);     // DHT 객체 생성
-float prevTemperature = 0.0;  // 이전 온도 값을 저장할 변수
+int prevTemperature = 0.0;  // 이전 온도 값을 저장할 변수
 
 void TemperatureSensor(String process, String lotId) {
   dht.begin();
-  float temperature = dht.readTemperature();  // 온도를 측정 (섭씨 ℃)
+  int temperature = dht.readTemperature();  // 온도를 측정 (섭씨 ℃)
 
   // 이전 온도 값과 새로운 온도 값이 다른 경우에만 SendTestData 함수 호출
   if (temperature != prevTemperature) {
